requirements.txt
cffi==1.11.5
Django==2.0.5
django-cleanup==2.1.0
django-widget-tweaks==1.4.2
Pillow==5.1.0
pycparser==2.18
pytz==2018.3
style==1.1.0
update==0.0.1
virtualenv==15.1.0
